+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 
 This file documents any relevant changes.
+## [0.9.10] - 2024-05.06
+- fix: Improve and refactor entrypoint to make --workers and --threads argument working #7
+
+## [0.9.9] - 2024-01.18
+- feat: added support timeout parameter
+- refactor: Change pip version from exact to minimum conditions
+
+## [0.9.8] - 2023-11.22
+- feat: added support for custom environment variables
+
+## [0.9.7] - 2023-10.27
+- fix: readded subdomain support
+- feat: added editorconf
+
 ## [0.9.6] - 2023-09-15
 - fix: updated grpcio to support python 3.11
 - fix: gunicorn will be patched on start up